backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
websockets==12.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
shap==0.44.1
joblib==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
geopy==2.4.1
psycopg2-binary==2.9.6